Add auth listener to app to prevent uncorrected behavioral
