+++ angularclient/src/app/uicomponents/header/header.component.html
@@ -1,4 +1,4 @@
-<div id="application-header" data-name="application-header-wrapper" *ngIf="userIsLoggedIn">
+<div id="application-header" data-name="application-header-wrapper" > <!--*ngIf="userIsLoggedIn"-->
   <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
     <a class="navbar-brand" href="#">SimpleCRM</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
@@ -27,8 +27,8 @@
         </li>
       </ul>
 
-      <span id="userName" class="nav-link currentUserName" *ngIf="currentLoggedInUserName">Hi: <span class="currentUserNameColor">{{currentLoggedInUserName}}</span></span>
-      <a class="nav-link" (click)="logout()">Logout <span class="sr-only">(current)</span></a>
+<!--      <span id="userName" class="nav-link currentUserName" *ngIf="currentLoggedInUserName">Hi: <span class="currentUserNameColor">{{currentLoggedInUserName}}</span></span>-->
+<!--      <a class="nav-link" (click)="logout()">Logout <span class="sr-only">(current)</span></a>-->
     </div>
 
   </nav>
